player can now play game almost fully
music from level01 plays when player transitions to level02. this was not intentional but gives off an eerie vibe to be in the middle of a sandstorm and search for a way out.

need to debug inventory bugs for the final scene.

add more things in the scenes for interest.